requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
supabase==2.6.0
httpx==0.27.2
